requirements.txt
boto==2.38.0
Jinja2==2.8
MarkupSafe==0.23
wheel==0.24.0
PyYAML==3.11
